feat(mcp): extract JSON-RPC handling from ToolRegistry to McpToolHandler
- Create McpToolHandler interface and DefaultMcpToolHandler implementation
- Move JSON-RPC protocol methods (handleToolsList, handleToolsCall) from ToolRegistry
- Remove formatErrorForJsonRpc method from ToolRegistry for clean domain separation
- Enhance JSON response handling with pretty-printing for complex objects
- Achieve complete separation of concerns: domain vs protocol responsibilities

Resolves SPI-632: JSON-RPC extraction for better separation of concerns
